requirements.txt
asgiref==3.9.0
Django==5.2.4
django-js-asset==3.1.2
django-mptt==0.17.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
